break out NodeBalance as its own component with its own css
