+++ satisfactory-accounting-app/src/app.scss
@@ -1,5 +1,5 @@
 @use "colors.scss";
-@use "node_display.scss";
+@use "node_display/node_display.scss";
 @use "clickedit/ClickEdit.scss";
 
 body {
