+++ satisfactory-accounting-app/src/node_display/building/building_type.rs
@@ -8,9 +8,9 @@
 use satisfactory_accounting::database::BuildingId;
 use yew::prelude::*;
 
+use crate::context::CtxHelper;
 use crate::node_display::building::choose_from_list::{Choice, ChooseFromList};
 use crate::node_display::icon::Icon;
-use crate::CtxHelper;
 
 #[derive(PartialEq, Properties)]
 pub struct Props {
