+++ satisfactory-accounting-app/src/node_display/building/item.rs
@@ -9,9 +9,9 @@ use log::warn;
 use satisfactory_accounting::database::{BuildingId, BuildingKind, ItemId};
 use yew::prelude::*;
 
+use crate::context::CtxHelper;
 use crate::node_display::building::choose_from_list::{Choice, ChooseFromList};
 use crate::node_display::icon::Icon;
-use crate::CtxHelper;
 
 #[derive(PartialEq, Properties)]
 pub struct Props {
