+++ satisfactory-accounting-app/src/node_display/node_display.scss
@@ -1,5 +1,7 @@
 @use "sass:color";
-@use "colors.scss";
+@use "../colors.scss";
+
+@use "balance/NodeBalance.scss";
 
 .NodeDisplay {
     .drag-handle {
@@ -184,79 +186,6 @@
             gap: 5px;
         }
 
-    }
-
-    .balance {
-        box-sizing: border-box;
-        display: flex;
-        justify-content: flex-start;
-        align-items: stretch;
-
-        .item-entries {
-            box-sizing: border-box;
-            display: flex;
-            margin: 0;
-            padding: 0;
-        }
-
-        .entry-row {
-            box-sizing: border-box;
-
-            display: flex;
-            flex-direction: row;
-            justify-content: space-between;
-            align-items: center;
-            cursor: pointer;
-
-            .balance-value {
-                text-align: right;
-            }
-
-            &.negative {
-                .balance-value {
-                    color: colors.$danger;
-                }
-            }
-            &.positive {
-                .balance-value {
-                    color: color.scale(colors.$success, $lightness: -30%);
-                }
-            }
-        }
-
-        &.vertical {
-            flex-direction: column;
-
-            border-radius: 5px;
-            background-color: colors.$light;
-            padding: 5px;
-            gap: 5px;
-
-            .item-entries {
-                flex-direction: column;
-                gap: 5px;
-            }
-
-            .entry-row {
-                .balance-value {
-                    min-width: 3em;
-                }
-            }
-        }
-
-        &.horizontal {
-            flex-direction: row-reverse;
-            gap: 10px;
-
-            .item-entries {
-                flex-direction: row;
-                gap: 10px;
-            }
-
-            .entry-row {
-                gap: 2px;
-            }
-        }
     }
 
     .ChooseFromList {
